--- src/Screens/HomeScreen.tsx
@@ -1,6 +1,0 @@
-import { ReactElement,  FC } from "react";
-
-const Link1:FC = ():ReactElement=>{
-    return(<h1>link1</h1>)
-}
-export default Link1;
