--- src/Screens/SignInScreen.tsx
@@ -1,6 +1,0 @@
-import { ReactElement,  FC } from "react";
-
-const Link2:FC = ():ReactElement=>{
-    return(<h1>link2</h1>)
-}
-export default Link2;
