--- src/Screens/SignUpScreen.tsx
@@ -1,6 +1,0 @@
-import { ReactElement,  FC } from "react";
-
-const Link3:FC = ():ReactElement=>{
-    return(<h1>link3</h1>)
-}
-export default Link3;
