+++ src/screens/Home.tsx
@@ -1,0 +1,6 @@
+import { ReactElement,  FC } from "react";
+
+const Link1:FC = ():ReactElement=>{
+    return(<h1>link1</h1>)
+}
+export default Link1;
