+++ src/screens/SignUp.tsx
@@ -1,0 +1,6 @@
+import { ReactElement,  FC } from "react";
+
+const Link3:FC = ():ReactElement=>{
+    return(<h1>link3</h1>)
+}
+export default Link3;
